requirements.txt
streamlit
openai
python-dotenv
openai-whisper
